# docs/swagger.yaml
definitions:
...
  cms_internal_app_role_internal_dto.ApiPagedData-dto_Role:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.Role'
        type: array
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  cms_internal_app_user_internal_dto.ApiPagedData-dto_User:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.User'
        type: array
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  dto.ApiRole:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.Role'
      msg:
        type: string
    type: object
  dto.ApiRoleList:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/cms_internal_app_role_internal_dto.ApiPagedData-dto_Role'
      msg:
        type: string
    type: object
  dto.ApiUser:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.User'
      msg:
        type: string
    type: object
  dto.ApiUserList:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/cms_internal_app_user_internal_dto.ApiPagedData-dto_User'
      msg:
        type: string
    type: object
  dto.ApiUserLogin:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.UserAuthnResult'
      msg:
        type: string
    type: object
...
  dto.Role:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.RoleCreateParams:
    properties:
      description:
        maxLength: 255
        type: string
      inheritList:
        items:
          type: string
        maxItems: 10
        type: array
      name:
        maxLength: 255
        type: string
    required:
    - description
    - name
    type: object
  dto.RoleListParams:
    properties:
      description:
        type: string
      name:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
    type: object
  dto.RoleUpdateParams:
    properties:
      description:
        maxLength: 255
        type: string
      id:
        example: "123456"
        type: string
      inheritList:
        items:
          type: string
        maxItems: 10
        type: array
      name:
        maxLength: 255
        type: string
    required:
    - description
    - id
    - name
    type: object
  dto.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      ip:
        type: string
      profile:
        $ref: '#/definitions/dto.UserProfile'
      roleId:
        type: string
      roleName:
        type: string
      status:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  dto.UserAuthnResult:
    properties:
      detail:
        $ref: '#/definitions/dto.User'
      token:
        type: string
    type: object
  dto.UserCreateParams:
    properties:
      email:
        maxLength: 64
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      passwordConfirm:
        maxLength: 32
        minLength: 6
        type: string
      username:
        maxLength: 32
        minLength: 4
        type: string
    required:
    - email
    - password
    - passwordConfirm
    - username
    type: object
  dto.UserCredentialParams:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UserListParams:
    properties:
      city:
        type: string
      country:
        type: string
      email:
        type: string
      gender:
        enum:
        - 0
        - 1
        type: integer
      ip:
        type: string
      nickname:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
      province:
        type: string
      roleId:
        type: string
      status:
        type: integer
      username:
        type: string
      verifyEndTime:
        type: string
      verifyStartTime:
        type: string
    type: object
  dto.UserPasswdResetParams:
    properties:
      id:
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - id
    - password
    type: object
  dto.UserPasswdUpdateParams:
    properties:
      oldPassword:
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      passwordConfirm:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - oldPassword
    - password
    - passwordConfirm
    type: object
  dto.UserProfile:
    properties:
      avatar:
        type: string
      city:
        type: string
      country:
        type: string
      description:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      province:
        type: string
    type: object
...
  types.ApiCreateResult:
    properties:
      code:
        type: string
      data:
        description: 数据 ID
        example: "123456"
        type: string
      msg:
        type: string
    type: object
  types.ApiEmptyResult:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
  types.ResourceID:
    properties:
      id:
        example: "123456"
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  license:
    name: Apache 2.0
  title: EgoCms Api
  version: "0.1"
paths:
...
  /role/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RoleCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiCreateResult'
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - 角色
      x-apifox-folder: 角色
  /role/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - 角色
      x-apifox-folder: 角色
  /role/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiRole'
      security:
      - ApiKeyAuth: []
      summary: 查看角色信息
      tags:
      - 角色
      x-apifox-folder: 角色
  /role/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RoleListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiRoleList'
      security:
      - ApiKeyAuth: []
      summary: 角色列表
      tags:
      - 角色
      x-apifox-folder: 角色
  /role/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RoleUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - 角色
      x-apifox-folder: 角色
  /user/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswdUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiUser'
      security:
      - ApiKeyAuth: []
      summary: 查看用户资料
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiUserList'
      security:
      - ApiKeyAuth: []
      summary: 查看用户列表
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserCredentialParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiUserLogin'
      summary: 用户登录
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 注销登录
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiCreateResult'
      summary: 用户注册
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswdResetParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 重置密码
      tags:
      - 用户
      x-apifox-folder: 用户
  /user/update-profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新用户资料
      tags:
      - 用户
      x-apifox-folder: 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"